Move get_ammo(), find_ammo(), find_reloadables() and ammo_count_for()
