+++ src/character_functions.cpp
@@ -11,6 +11,7 @@
 #include "itype.h"
 #include "make_static.h"
 #include "map_iterator.h"
+#include "map_selector.h"
 #include "messages.h"
 #include "monster.h"
 #include "npc.h"
@@ -22,6 +23,7 @@
 #include "uistate.h"
 #include "veh_type.h"
 #include "vehicle.h"
+#include "vehicle_selector.h"
 #include "vpart_position.h"
 #include "weather_gen.h"
 #include "weather.h"
@@ -50,6 +52,7 @@ static const trait_id trait_WEB_WEAVER( "WEB_WEAVER" );
 
 static const std::string flag_FUNGUS( "FUNGUS" );
 static const std::string flag_SWIMMABLE( "SWIMMABLE" );
+static const std::string flag_SPEEDLOADER( "SPEEDLOADER" );
 
 static const efftype_id effect_boomered( "boomered" );
 static const efftype_id effect_darkness( "darkness" );
@@ -60,6 +63,7 @@ static const bionic_id bio_uncanny_dodge( "bio_uncanny_dodge" );
 
 static const itype_id itype_battery( "battery" );
 static const itype_id itype_cookbook_human( "cookbook_human" );
+static const itype_id itype_UPS( "UPS" );
 
 namespace character_funcs
 {
@@ -717,7 +721,8 @@ bool list_ammo( const Character &who, const item &base, std::vector<item::reload
     bool ammo_match_found = false;
     int ammo_search_range = who.is_mounted() ? -1 : 1;
     for( const auto e : opts ) {
-        for( const item_location &ammo : who.find_ammo( *e, include_empty_mags, ammo_search_range ) ) {
+        for( const item_location &ammo : find_ammo_items_or_mags( who, *e, include_empty_mags,
+                ammo_search_range ) ) {
             // don't try to unload frozen liquids
             if( ammo->is_watertight_container() && ammo->contents_made_of( SOLID ) ) {
                 continue;
@@ -1032,6 +1037,184 @@ item::reload_option select_ammo( const Character &who, const item &base, bool pr
     }
 
     return select_ammo( who, base, std::move( ammo_list ) );
+}
+
+std::vector<const item *> get_ammo_items( const Character &who, const ammotype &at )
+{
+    return who.items_with( [at]( const item & it ) {
+        return it.ammo_type() == at;
+    } );
+}
+
+template <typename T, typename Output>
+void find_ammo_helper( T &src, const item &obj, bool empty, Output out, bool nested )
+{
+    if( obj.is_watertight_container() ) {
+        if( !obj.is_container_empty() ) {
+            auto contents_id = obj.contents.front().typeId();
+
+            // Look for containers with the same type of liquid as that already in our container
+            src.visit_items( [&src, &nested, &out, &contents_id, &obj]( item * node ) {
+                if( node == &obj ) {
+                    // This stops containers and magazines counting *themselves* as ammo sources.
+                    return VisitResponse::SKIP;
+                }
+
+                if( node->is_container() && !node->is_container_empty() &&
+                    node->contents.front().typeId() == contents_id ) {
+                    out = item_location( src, node );
+                }
+                return nested ? VisitResponse::NEXT : VisitResponse::SKIP;
+            } );
+        } else {
+            // Look for containers with any liquid
+            src.visit_items( [&src, &nested, &out]( item * node ) {
+                if( node->is_container() && node->contents_made_of( LIQUID ) ) {
+                    out = item_location( src, node );
+                }
+                return nested ? VisitResponse::NEXT : VisitResponse::SKIP;
+            } );
+        }
+    }
+    if( obj.magazine_integral() ) {
+        // find suitable ammo excluding that already loaded in magazines
+        const std::set<ammotype> &ammo = obj.ammo_types();
+        const auto mags = obj.magazine_compatible();
+
+        src.visit_items( [&src, &nested, &out, &mags, ammo]( item * node ) {
+            if( node->is_gun() || node->is_tool() ) {
+                // guns/tools never contain usable ammo so most efficient to skip them now
+                return VisitResponse::SKIP;
+            }
+            if( !node->made_of( SOLID ) ) {
+                // some liquids are ammo but we can't reload with them unless within a container or frozen
+                return VisitResponse::SKIP;
+            }
+            if( node->is_ammo_container() && !node->contents.empty() &&
+                !node->contents_made_of( SOLID ) ) {
+                for( const ammotype &at : ammo ) {
+                    if( node->contents.front().ammo_type() == at ) {
+                        out = item_location( src, node );
+                    }
+                }
+                return VisitResponse::SKIP;
+            }
+
+            for( const ammotype &at : ammo ) {
+                if( node->ammo_type() == at ) {
+                    out = item_location( src, node );
+                }
+            }
+            if( node->is_magazine() && node->has_flag( flag_SPEEDLOADER ) ) {
+                if( mags.count( node->typeId() ) && node->ammo_remaining() ) {
+                    out = item_location( src, node );
+                }
+            }
+            return nested ? VisitResponse::NEXT : VisitResponse::SKIP;
+        } );
+    } else {
+        // find compatible magazines excluding those already loaded in tools/guns
+        const auto mags = obj.magazine_compatible();
+
+        src.visit_items( [&src, &nested, &out, mags, empty]( item * node ) {
+            if( node->is_gun() || node->is_tool() ) {
+                return VisitResponse::SKIP;
+            }
+            if( node->is_magazine() ) {
+                if( mags.count( node->typeId() ) && ( node->ammo_remaining() || empty ) ) {
+                    out = item_location( src, node );
+                }
+                return VisitResponse::SKIP;
+            }
+            return nested ? VisitResponse::NEXT : VisitResponse::SKIP;
+        } );
+    }
+}
+
+std::vector<item_location> find_ammo_items_or_mags( const Character &who, const item &obj, bool empty,
+        int radius )
+{
+    std::vector<item_location> res;
+
+    find_ammo_helper( const_cast<Character &>( who ), obj, empty, std::back_inserter( res ), true );
+
+    if( radius >= 0 ) {
+        for( auto &cursor : map_selector( who.pos(), radius ) ) {
+            find_ammo_helper( cursor, obj, empty, std::back_inserter( res ), false );
+        }
+        for( auto &cursor : vehicle_selector( who.pos(), radius ) ) {
+            find_ammo_helper( cursor, obj, empty, std::back_inserter( res ), false );
+        }
+    }
+
+    return res;
+}
+
+std::vector<item_location> find_reloadables( const Character &who )
+{
+    std::vector<item_location> reloadables;
+
+    who.visit_items( [&]( const item * node ) {
+        if( node->is_holster() ) {
+            return VisitResponse::NEXT;
+        }
+        bool reloadable = false;
+        if( node->is_gun() && !node->magazine_compatible().empty() ) {
+            reloadable = node->magazine_current() == nullptr ||
+                         node->ammo_remaining() < node->ammo_capacity();
+        } else {
+            reloadable = ( node->is_magazine() || node->is_bandolier() ||
+                           ( node->is_gun() && node->magazine_integral() ) ) &&
+                         node->ammo_remaining() < node->ammo_capacity();
+        }
+        if( reloadable ) {
+            reloadables.push_back( item_location( const_cast<Character &>( who ),
+                                                  const_cast<item *>( node ) ) );
+        }
+        return VisitResponse::SKIP;
+    } );
+    return reloadables;
+}
+
+int ammo_count_for( const Character &who, const item &gun )
+{
+    int ret = item::INFINITE_CHARGES;
+    if( !gun.is_gun() ) {
+        return ret;
+    }
+
+    int required = gun.ammo_required();
+
+    if( required > 0 ) {
+        int total_ammo = 0;
+        total_ammo += gun.ammo_remaining();
+
+        bool has_mag = gun.magazine_integral();
+
+        const auto found_ammo = find_ammo_items_or_mags( who, gun, true, -1 );
+        int loose_ammo = 0;
+        for( const auto &ammo : found_ammo ) {
+            if( ammo->is_magazine() ) {
+                has_mag = true;
+                total_ammo += ammo->ammo_remaining();
+            } else if( ammo->is_ammo() ) {
+                loose_ammo += ammo->charges;
+            }
+        }
+
+        if( has_mag ) {
+            total_ammo += loose_ammo;
+        }
+
+        ret = std::min( ret, total_ammo / required );
+    }
+
+    int ups_drain = gun.get_gun_ups_drain();
+    if( ups_drain > 0 ) {
+        ret = std::min( ret, who.charges_of( itype_UPS ) / ups_drain );
+    }
+
+    return ret;
 }
 
 } // namespace character_funcs
